+++ protocol/x/clob/keeper/grpc_stream_orderbook.go
@@ -8,7 +8,10 @@ func (k Keeper) StreamOrderbookUpdates(
 	req *types.StreamOrderbookUpdatesRequest,
 	stream types.Query_StreamOrderbookUpdatesServer,
 ) error {
-	err := k.GetFullNodeStreamingManager().Subscribe(*req, stream)
+	err := k.GetFullNodeStreamingManager().Subscribe(
+		req.GetClobPairId(),
+		stream,
+	)
 	if err != nil {
 		return err
 	}
